@@ -1,23 +1,23 @@
-const bootstrapIcon = require("./assets/Images/bootstrap.png");
-const htmlIcon = require("./assets/Images/html-5.png");
-const javascriptIcon = require("./assets/Images/javascript.png");
-const cssIcon = require("./assets/Images/css.png");
-const mongoDBIcon = require("./assets/Images/mongo-db.png");
-const nodeJSIcon = require("./assets/Images/nodejs.png");
-const reactIcon = require("./assets/Images/react.png");
-const sassIcon = require("./assets/Images/sass.png");
-const pythonIcon = require("./assets/Images/python.png");
-const figmaIcon = require("./assets/Images/figma.png");
-const herokuIcon = require("./assets/Images/heroku.png");
-const gitIcon = require("./assets/Images/git.png");
-const githubIcon = require("./assets/Images/github.png");
-const postgreIcon = require("./assets/Images/postgre.png");
-const tsIcon = require("./assets/Images/typescript.png");
-const slackIcon = require("./assets/Images/slack.png");
-const jiraIcon = require("./assets/Images/jira.png");
-const tailwindIcon = require("./assets/Images/tailwind.png");
-const jestIcon = require("./assets/Images/jest.png");
-const nextIcon = require("./assets/Images/next.png");
+import bootstrapIcon from "../assets/Images/bootstrap.png";
+import htmlIcon from "../assets/Images/html-5.png";
+import javascriptIcon from "../assets/Images/javascript.png";
+import cssIcon from "../assets/Images/css.png";
+import mongoDBIcon from "../assets/Images/mongo-db.png";
+import nodeJSIcon from "../assets/Images/nodejs.png";
+import reactIcon from "../assets/Images/react.png";
+import sassIcon from "../assets/Images/sass.png";
+import pythonIcon from "../assets/Images/python.png";
+import figmaIcon from "../assets/Images/figma.png";
+import herokuIcon from "../assets/Images/heroku.png";
+import gitIcon from "../assets/Images/git.png";
+import githubIcon from "../assets/Images/github.png";
+import postgreIcon from "../assets/Images/postgre.png";
+import tsIcon from "../assets/Images/typescript.png";
+import slackIcon from "../assets/Images/slack.png";
+import jiraIcon from "../assets/Images/jira.png";
+import tailwindIcon from "../assets/Images/tailwind.png";
+import jestIcon from "../assets/Images/jest.png";
+import nextIcon from "../assets/Images/next.png";
 
 export const SkillInfo = [
   {
