@@ -2,9 +2,9 @@ import React from "react";
 
 export default function ThoughtBubble() {
   return (
-    <div class="thought-bubble">
-      <div class="thought-bubble__point"></div>
-      <p class="thought-bubble__txt">
+    <div className="thought-bubble">
+      <div className="thought-bubble__point"></div>
+      <p className="thought-bubble__txt">
         "I wonder what would happen if someone were to click my face?"
       </p>
     </div>
